gui: Add header file for the Autonumber dialog
