+++ libleptongui/src/x_autonumber.c
@@ -40,78 +40,6 @@
 /** @brief How many entries to keep in the "Search text" combo box. */
 #define HISTORY_LENGTH 15
 
-/* autonumber text structs and enums */
-enum {
-  AUTONUMBER_SORT_DIAGONAL,
-  AUTONUMBER_SORT_YX,
-  AUTONUMBER_SORT_YX_REV,
-  AUTONUMBER_SORT_XY,
-  AUTONUMBER_SORT_XY_REV,
-  AUTONUMBER_SORT_FILE
-};
-
-enum {
-  AUTONUMBER_IGNORE,
-  AUTONUMBER_RENUMBER,
-  AUTONUMBER_RESPECT
-};
-
-enum {
-  SCOPE_SELECTED,
-  SCOPE_PAGE,
-  SCOPE_HIERARCHY
-};
-
-typedef struct autonumber_text_t AUTONUMBER_TEXT;
-
-/** @brief Stored state of the autonumber text dialog */
-struct autonumber_text_t {
-  /** @brief Search text history */
-  GList *scope_text;
-
-  /** @brief Scope for searching existing numbers */
-  gint scope_skip;
-
-  /** @brief Scope for autonumbering text */
-  gint scope_number;
-
-  /** @brief Overwrite existing numbers in scope */
-  gboolean scope_overwrite;
-
-  /** @brief Sort order */
-  gint order;
-
-  /** @brief Starting number for automatic numbering */
-  gint startnum;
-
-  /** @brief Remove numbers instead of automatic numbering */
-  gboolean removenum;
-
-  /** @brief Automatic assignments of slots */
-  gboolean slotting;
-
-  /** @brief Pointer to the dialog */
-  GtkWidget *dialog;
-
-  /** @brief Pointer to the SchematicWindow struct */
-  SchematicWindow *w_current;
-
-  /* variables used while autonumbering */
-  gchar * current_searchtext;
-  gint root_page;      /* flag whether its the root page or not */
-  GList *used_numbers; /* list of used numbers */
-  GList *free_slots;   /* list of FREE_SLOT objects */
-  GList *used_slots;   /* list of USED_SLOT objects */
-};
-
-typedef struct autonumber_slot_t AUTONUMBER_SLOT;
-
-struct autonumber_slot_t {
-  gchar *symbolname;     /* or should I use the device name? (Werner) */
-  gint number;           /* usually this is the refdes number */
-  gint slotnr;      /* just the number of the free slot */
-};
-
 
 static AUTONUMBER_TEXT *autotext = NULL;
